Modify duplicate movies logic
I'm now selecting duplicates by looking at their titles only, so
this brings more hits, some of which are false positives.

This is ok, because for example Get Carter (1972)/Get Carter (1994)
are different movies.

So I added a field and corresponding UI to declare a movie as not
duplicate, so that it won't show up as a duplicate anymore.

Related to #16